PRJ_21_46309_48594_49154/03_Implementacao/talkID/app_comm/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
dj-database-url==2.2.0
Django==4.2.13
django-pwa==1.1.0
djangorestframework==3.15.2
frozenlist==1.4.1
http-ece==1.2.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
multidict==6.0.5
psycopg2==2.9.9
py-vapid==1.9.1
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-decouple==3.8
pytz==2024.1
pywebpush==2.0.0
redis==5.0.7
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
twisted==24.3.0
txaio==23.1.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
yarl==1.9.4
zope.interface==6.4.post2
